Kotlin Android Tutorial - New Eat It part 16 Load Cart Items

+++ app/src/main/java/com/createsapp/kotlineatitv2client/database/CartDataSource.kt
@@ -9,7 +9,7 @@ interface CartDataSource {
 
     fun countItemInCart(uid: String): Single<Int>
 
-    fun sumPrice(uid: String): Single<Long>
+    fun sumPrice(uid: String): Single<Double>
 
     fun getItemInCart(foodId: String, uid: String): Single<CartItem>
